pyproject.toml
[tool.poetry]
name = "layer-3"
version = "0.1.0"
description = ""
authors = ["Matt Sárdi <sardi.matt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.18"
pyautogui = "^0.9.54"
pillow = "^10.1.0"
pyprojroot = "^0.3.0"
torch = "^2.1.0"
accelerate = "^0.24.1"
salesforce-lavis = "^1.0.2"
transformers = "^4.26"
pysimplegui = "^4.60.5"
jupyter = "^1.0.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
